restructure entry points
