+++ microservices/backTesting/go.mod
@@ -19,6 +19,7 @@ require (
 
 require (
 	github.com/google/uuid v1.6.0 // indirect
+	github.com/joho/godotenv v1.5.1 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/vektah/gqlparser/v2 v2.5.19 // indirect
