+++ microservices/backTesting/main.go
@@ -1,25 +1,22 @@
 package main
 
 import (
-	"context"
 	"flag"
 	"fmt"
-	"io/fs"
-	"net/http"
 	"os"
-	"path/filepath"
-	"sort"
-	"strings"
+
+	"github.com/joho/godotenv"
 
 	"cryptobotmanager.com/cbm-backend/microservices/backTesting/functions"
-	"cryptobotmanager.com/cbm-backend/resolvers/graph/model"
 	"cryptobotmanager.com/cbm-backend/shared"
-	"github.com/Khan/genqlient/graphql"
-	"github.com/rs/zerolog/log"
-	// Add any other necessary imports
+	sharedlog "github.com/rs/zerolog/log"
 )
 
 func main() {
+	err := godotenv.Load(".env")
+	if err != nil {
+		fmt.Println("Warning: No .env file found or failed to load")
+	}
 
 	// Initialize logger
 	shared.SetupLogger()
@@ -41,53 +38,20 @@ func main() {
 		backend = "http://resolvers:8080/query"
 	}
 
-	client := graphql.NewClient(backend, &http.Client{})
-	ctx := context.Background()
+	fmt.Println("SYSTEM_MODE is:", os.Getenv("SYSTEM_MODE"))
 
-	// Directory with your JSON price files
-	dataDir := "../../binancePrices"
-
-	// Read and sort files chronologically
-	var files []string
-	filepath.WalkDir(dataDir, func(path string, d fs.DirEntry, err error) error {
+	// Setup GraphQL backend client
+	if os.Getenv("SYSTEM_MODE") == "local" {
+		err := functions.CSVPrices(backend)
 		if err != nil {
-			log.Error().Err(err).Msg("Error walking through files")
-			return err
+			sharedlog.Error().Err(err).Msg("Failed to load JSON data")
 		}
-		if !d.IsDir() && strings.HasPrefix(d.Name(), "binance_prices_") && strings.HasSuffix(d.Name(), ".json") {
-			files = append(files, path)
-		}
-		return nil
-	})
-	sort.Strings(files) // Ensure they’re processed in date order
 
-	// Loop through each file and "replay" the data
-	for _, file := range files {
-		log.Info().Str("File", file).Msg("Processing file")
-
-		marketData, err := functions.LoadPriceSnapshotsFromFile(file)
+	} else {
+		err := functions.BinancePrices(backend)
 		if err != nil {
-			log.Error().Err(err).Str("file", file).Msg("Failed to load JSON data")
-			continue
+			sharedlog.Error().Err(err).Msgf("Failed to get price data from Binance!")
 		}
-
-		for _, snapshot := range marketData {
-			// Convert []Price (in snapshot.Pairs) to []PriceData
-			var market []model.Pair
-			for _, p := range snapshot.Pair {
-				market = append(market, model.Pair{
-					Symbol: p.Symbol,
-					Price:  p.Price,
-				})
-			}
-
-			err := functions.SavePriceData(ctx, client, market, int(snapshot.Timestamp))
-			if err != nil {
-				log.Error().Err(err).Int("timestamp", snapshot.Timestamp).Msg("Failed to save snapshot")
-			} else {
-				log.Info().Int("timestamp", snapshot.Timestamp).Msg("Replayed snapshot")
-			}
-		}
-
 	}
+
 }
