+++ shared/structs.go
@@ -1,0 +1,1 @@
+package shared
